moved ExchangeChartEntity to new AggregationResults abstraction
